E.2: migrate 10 per-command helpers from _cmd_helpers.py into cmds/*.py
Each helper method is now a free function in the ``cmds/*.py`` file
that owns its command.  A tiny ``_Cmd<Group>Mixin`` class in each file
exposes the functions as methods via assignment (Python's descriptor
protocol binds ``self`` as the first arg), so every external caller
keeps working unchanged:

  cmds/return_.py    → _emit_cmd_return           (+ _CmdReturnMixin)
  cmds/array_.py     → _emit_array_subcmd_value   (new file)
                       _emit_array_set_list        (+ _CmdArrayMixin)
  cmds/uplevel_.py   → _emit_unset_array_elems
                       _emit_cmd_uplevel
                       _emit_uplevel_body          (+ _CmdUplevelMixin)
  cmds/list_.py      → _emit_list_value
                       _emit_cmd_lassign           (+ _CmdListMixin)
  cmds/clock_.py     → _emit_clock_value           (new file + _CmdClockMixin)
  cmds/info_.py      → _emit_info_value            (+ _CmdInfoMixin)

_emitter/__init__.py now composes _WasmEmitter from the 6 per-command
mixins alongside the existing generic mixins.  ``_cmd_helpers.py`` is
deleted — every per-command Python codegen now lives in the same file
as its BUILTINS-side hook.

Adding a new Tcl command's Python codegen is one file: create or
extend cmds/<cmd>_.py with the hook + helpers + mixin class; add the
mixin to _emitter/__init__.py's composition.

Runtime unchanged; 787 WASM tests pass; parity baseline unchanged.
